Move benchmarks outside core
