+++ evolvekit/benchmarks/__init__.py
@@ -1,0 +1,10 @@
+# GA benchmarks
+from evolvekit.benchmarks.SphereEvaluator import SphereEvaluator
+from evolvekit.benchmarks.RastriginEvaluator import RastriginEvaluator
+from evolvekit.benchmarks.RosenbrockEvaluator import RosenbrockEvaluator
+
+__all__ = [
+    "SphereEvaluator",
+    "RastriginEvaluator",
+    "RosenbrockEvaluator",
+]
