+++ evolvekit/core/__init__.py
@@ -1,9 +1,7 @@
 # EvolveKit Core Components
 from evolvekit.core.Ga import *
-from evolvekit.core.benchmarks import *
 
 # Combine __all__
 from evolvekit.core.Ga import __all__ as ga_all
-from evolvekit.core.benchmarks import __all__ as benchmarks_all
 
-__all__ = ga_all + benchmarks_all
+__all__ = ga_all
