--- evolvekit/core/benchmarks/__init__.py
@@ -1,10 +1,0 @@
-# GA benchmarks
-from evolvekit.core.benchmarks.SphereEvaluator import SphereEvaluator
-from evolvekit.core.benchmarks.RastriginEvaluator import RastriginEvaluator
-from evolvekit.core.benchmarks.RosenbrockEvaluator import RosenbrockEvaluator
-
-__all__ = [
-    "SphereEvaluator",
-    "RastriginEvaluator",
-    "RosenbrockEvaluator",
-]
