+++ evolvekit/examples/advanced/compare_clamp_strategies.py
@@ -1,6 +1,6 @@
 from evolvekit.core.Ga.GaIsland import GaIsland
 from evolvekit.core.Ga.enums.GaClampStrategy import GaClampStrategy
-from evolvekit.core.benchmarks import SphereEvaluator
+from evolvekit.benchmarks import SphereEvaluator
 
 from evolvekit.operators.Ga.selection.RankSelection import RankSelection
 from evolvekit.operators.Ga.crossover.real.OnePointCrossover import OnePointCrossover
